# api-specification.yaml
openapi: 3.0.3
info:
  title: StrellerMinds Backend API
  description: Minimal Blockchain Education Platform API
  version: 1.0.0
  contact:
    name: StarkMindsHQ
    url: https://github.com/StarkMindsHQ/StrellerMinds-Backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.strellerminds.com
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates a user and returns access tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Registers a new user account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Refreshes an access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Retrieves the authenticated user's profile
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /courses:
    get:
      tags:
        - Courses
      summary: Get all courses
      description: Retrieves a list of all courses
      operationId: getAllCourses
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: category
          in: query
          description: Filter by course category
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          description: Filter by difficulty level
          required: false
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoursesListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Courses
      summary: Create course
      description: Creates a new course
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /courses/{id}:
    get:
      tags:
        - Courses
      summary: Get course by ID
      description: Retrieves a specific course by its ID
      operationId: getCourseById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Course ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Courses
      summary: Update course
      description: Updates a course's information
      operationId: updateCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Course ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseRequest'
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Courses
      summary: Delete course
      description: Deletes a course
      operationId: deleteCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Course ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Course deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
    paypalSignature:
      type: apiKey
      in: header
      name: paypal-transmission-sig
    zoomSignature:
      type: apiKey
      in: header
      name: x-zm-signature
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: Request path
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
        rememberMe:
          type: boolean
          description: Remember user session
          default: false
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          description: User first name
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          description: User last name
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        expiresIn:
          type: number
          description: Token expiration time in seconds
        tokenType:
          type: string
          enum: [Bearer]
          example: Bearer
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        email:
          type: string
          format: email
          description: User email
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        role:
          type: string
          enum: [user, admin, instructor]
          description: User role
        isActive:
          type: boolean
          description: Whether user account is active
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    UserProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        email:
          type: string
          format: email
          description: User email
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        role:
          type: string
          enum: [user, admin, instructor]
          description: User role
        profile:
          type: object
          properties:
            bio:
              type: string
              description: User bio
            avatar:
              type: string
              format: uri
              description: Avatar URL
            preferences:
              type: object
              description: User preferences
        enrollments:
          type: array
          items:
            $ref: '#/components/schemas/EnrollmentResponse'
          description: User course enrollments
...
    CourseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Course ID
        title:
          type: string
          description: Course title
        description:
          type: string
          description: Course description
        category:
          type: string
          description: Course category
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
          description: Course difficulty level
        price:
          type: number
          format: float
          description: Course price
        duration:
          type: number
          description: Course duration in hours
        instructor:
          $ref: '#/components/schemas/UserResponse'
        enrollmentCount:
          type: number
          description: Number of enrolled students
        isPublished:
          type: boolean
          description: Whether course is published
        createdAt:
          type: string
          format: date-time
          description: Course creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    CreateCourseRequest:
      type: object
      required:
        - title
        - description
        - category
        - difficulty
        - price
        - duration
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Course title
        description:
          type: string
          minLength: 1
          description: Course description
        category:
          type: string
          description: Course category
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
          description: Course difficulty level
        price:
          type: number
          format: float
          minimum: 0
          description: Course price
        duration:
          type: number
          minimum: 1
          description: Course duration in hours
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Course title
        description:
          type: string
          description: Course description
        category:
          type: string
          description: Course category
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
          description: Course difficulty level
        price:
          type: number
          format: float
          minimum: 0
          description: Course price
        duration:
          type: number
          minimum: 1
          description: Course duration in hours
        isPublished:
          type: boolean
          description: Whether course is published
    CoursesListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/CourseResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        total:
          type: number
          description: Total number of courses
    EnrollmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Enrollment ID
        userId:
          type: string
          format: uuid
          description: User ID
        courseId:
          type: string
          format: uuid
          description: Course ID
        status:
          type: string
          enum: [active, completed, dropped]
          description: Enrollment status
        enrolledAt:
          type: string
          format: date-time
          description: Enrollment timestamp
        completedAt:
          type: string
          format: date-time
          description: Completion timestamp
        progress:
          type: number
          minimum: 0
          maximum: 100
          description: Course completion percentage
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Items per page
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there are more pages
        hasPrev:
          type: boolean
          description: Whether there are previous pages